+++ services/marketplace/order/types.go
@@ -11,7 +11,7 @@ type MarketplaceOrderCreateRequest struct {
 
 	CallbackUrl *string `json:"callback_url,omitempty" tfsdk:"callback_url"`
 
-	Limits map[string]int64 `json:"limits,omitempty" tfsdk:"limits"`
+	Limits map[string]float64 `json:"limits,omitempty" tfsdk:"limits"`
 
 	Offering *string `json:"offering" tfsdk:"offering"`
 
@@ -35,7 +35,7 @@ type MarketplaceOrderCreateLimitsRequest struct {
 type MarketplaceOrderUpdateRequest struct {
 	Attributes map[string]interface{} `json:"attributes,omitempty" tfsdk:"attributes"`
 
-	Limits map[string]int64 `json:"limits,omitempty" tfsdk:"limits"`
+	Limits map[string]float64 `json:"limits,omitempty" tfsdk:"limits"`
 
 	StartDate *string `json:"start_date,omitempty" tfsdk:"start_date"`
 }
@@ -119,11 +119,13 @@ type MarketplaceOrderResponse struct {
 
 	Issue *MarketplaceOrderIssueResponse `json:"issue,omitempty" tfsdk:"issue"`
 
-	Limits map[string]int64 `json:"limits,omitempty" tfsdk:"limits"`
+	Limits map[string]float64 `json:"limits,omitempty" tfsdk:"limits"`
 
 	MarketplaceResourceUuid *string `json:"marketplace_resource_uuid,omitempty" tfsdk:"marketplace_resource_uuid"`
 
 	NewCostEstimate *string `json:"new_cost_estimate,omitempty" tfsdk:"new_cost_estimate"`
+
+	NewPlanBillingMode *string `json:"new_plan_billing_mode,omitempty" tfsdk:"new_plan_billing_mode"`
 
 	NewPlanName *string `json:"new_plan_name,omitempty" tfsdk:"new_plan_name"`
 
@@ -148,6 +150,8 @@ type MarketplaceOrderResponse struct {
 	OfferingUuid *string `json:"offering_uuid,omitempty" tfsdk:"offering_uuid"`
 
 	OldCostEstimate common.FlexibleNumber `json:"old_cost_estimate,omitempty" tfsdk:"old_cost_estimate"`
+
+	OldPlanBillingMode *string `json:"old_plan_billing_mode,omitempty" tfsdk:"old_plan_billing_mode"`
 
 	OldPlanName *string `json:"old_plan_name,omitempty" tfsdk:"old_plan_name"`